linked-list: replace lib.rs with skeleton interface
move implementation to example.rs
